requirements.txt
fastapi[all]

fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
